[LegalizeTypes][VP] Widen fixed length vectors to VP nodes
If we have a node with an illegal fixed length vector result type that
needs widened, e.g. `x:v6i32 = add a, b`
Instead of just widening it to: `x:v8i32 = add a, b`
We can widen it to the equivalent VP operation and set the EVL to the
exact number of elements needed: `x:v8i32 = vp_add a, b, mask=true, evl=6`
Provided that the target supports the equivalent VP operation on the
widened type.

This patch applies this technique when widening unary, binary and
ternary ops, as well as selects, loads and stores, but there are more
places this could be applied.

For loads and stores, scalable vectors are already widened using VP ops,
so there the existing logic is just reused for fixed length vectors.

Differential Revision: https://reviews.llvm.org/D148523