[TTI] Use users of GEP to guess access type in getGEPCost
Currently getGEPCost uses the target type of the GEP as a heuristic for
the type that will be accessed, to pass onto isLegalAddressingMode.
Targets use this to work out if a GEP can then be folded into the
load/store instruction that uses the GEP.
For example, on RISC-V loads and stores can have an offset added to a
base register folded into a single instruction, so the following GEP is
free:

%p = getelementptr i32, ptr %base, i32 42       ; getInstructionCost = 0
%x = load i32, ptr %p                           ; getInstructionCost = 1
------------------------------------------------------------------------
lw t0, a0(42)

However vector loads and stores cannot have an offset folded into them,
so the following GEP is costed:

%p = getelementptr <2 x i32>, ptr %base, i32 42 ; getInstructionCost = 1
%x = load <2 x i32>, ptr %p                     ; getInstructionCost = 1
------------------------------------------------------------------------
addi  a0, 42
vle32 v8, (a0)

The issue arises whenever there is a mismatch between the target type of
the GEP and the type that is actually accessed:

%p = getelementptr i32, ptr %base, i32 42       ; getInstructionCost = 0
%x = load <2 x i32>, ptr %p                     ; getInstructionCost = 1
------------------------------------------------------------------------
addi  a0, 42
vle32 v8, (a0)

Even though this GEP will result in an add instruction, because TTI
thinks it's loading an i32, it will think it can be folded and not
charge for it.

The target type can become mismatched with the memory access during
transformations, noticeably during SLP where a scalar base pointer will
be reused to perform a vector load or store.

This patch changes the interface of getGEPCost and its users to
explicitly pass in the users of the GEP, so that it can more accurately
determine the types of the memory accesses occurring.

Note that although this API takes in a list of users, in
getInstructionCost and other places we only pass in the first user to
avoid O(N) compile-time performance.

Test cases that test the cost of a GEP or rely on the cost of a GEP have
been updated to include dummy users of the instructions, as a GEP on its
own with no users is no longer charged for.

It also introduces a check for GEPs that have all zero indices, and
costs them as free. This would ideally be included in a separate patch,
but the regression doesn't manifest itself until the changes in this
patch are made: See https://reviews.llvm.org/D38085#inline-332210

Differential Revision: https://reviews.llvm.org/D149889