[RISCV] Set Fast flag for unaligned scalar memory accesses
The +unaligned-scalar-mem feature was added in
https://reviews.llvm.org/D126085 to allow subtargets to indicate that
they supported misaligned loads/stores with "sufficient" performance.
This is separate from whether or not the target actually supports
misaligned accesses, which could be determined from Zicclsm.

This patch then enables the Fast flag under the assumption that any subtarget
that declares support for +unaligned-scalar-mem will want to opt into
optimisations that take advantrage of misaligned scalar accesses, such as store
merging.

Differential Revision: https://reviews.llvm.org/D150771